# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Book Store API Documentation",
    "description": "A simple API to create, read, update and delete books. Readers can view and rate books.",
    "version": "1.0.0"
  },
  "host": "auto-mart-api.herokuapp.com",
  "basePath": "/api/v1",
  "schemes": [
    "https"
  ],
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Books"
    },
    {
      "name": "Rating"
    }
  ],
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "The user id."
        },
        "first_name": {
          "type": "string",
          "description": "The user first name."
        },
        "last_name": {
          "type": "string",
          "description": "The user last name."
        },
        "books": {
          "properties": {
            "items": {
              "type": "array",
              "items": {}
            }
          },
          "description": "All books belonging to the user."
        },
        "email": {
          "type": "string",
          "description": "The user email."
        },
        "password": {
          "type": "string",
          "description": "The user password."
        },
        "joined": {
          "type": "string",
          "format": "date-time",
          "description": "The user registration date."
        }
      }
    },
    "Book": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "The book Id."
        },
        "title": {
          "type": "string",
          "description": "The book title."
        },
        "description": {
          "type": "string",
          "description": "The book description."
        },
        "author": {
          "type": "string",
          "description": "The book author."
        },
        "publisher": {
          "type": "string",
          "description": "The book publisher"
        },
        "genre": {
          "type": "number",
          "description": "The genre of the book."
        },
        "posted_by": {
          "type": "string",
          "description": "The name of the user who added the book."
        },
        "poster_id": {
          "type": "string",
          "description": "The id of the user who listed the book."
        },
        "image": {
          "properties": {
            "items": {
              "type": "array",
              "items": {}
            }
          },
          "description": "The book image details containing the image url and id."
        },
        "ratingAvg": {
          "type": "string",
          "description": "The book average rating."
        },
        "date": {
          "type": "string",
          "format": "date-time",
          "description": "The listing date of the book."
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "responses": {
    "UnauthorizedError": {
      "description": "API token is missing or invalid. Please login.",
      "headers": {
        "Authorization": {
          "type": "string"
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "Registers a new user.",
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "User Information",
            "schema": {
              "type": "object",
              "required": [
                "first_name",
                "last_name",
                "email",
                "password",
                "confirm_password"
              ],
              "properties": {
                "first_name": {
                  "type": "string"
                },
                "last_name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "confirm_password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "description": "Login a user.",
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Login Credentials",
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/book": {
      "get": {
        "description": "View all books.",
        "tags": [
          "Books"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        }
      }
    },
    "/book/{book_id}": {
      "get": {
        "description": "View a specific book matching the id.",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "description": "5da1b7fc6c649900174252ea",
            "type": "string",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        }
      }
    },
    "/book/": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Adds a new book to the store.",
        "tags": [
          "Books"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "title",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "author",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "publisher",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "genre",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "401": {
            "$ref": "#/responses/UnauthorizedError"
          }
        }
      }
    },
    "/book/{book_id}/": {
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Updates a listed book.",
        "tags": [
          "Books"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "description": "5da25a1d0fa5bf00171f96d2",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "Book Information",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "author": {
                  "type": "string"
                },
                "publisher": {
                  "type": "string"
                },
                "genre": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/book/{book_id}/delete": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Deletes a book.",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "description": "5da25a1d0fa5bf00171f96d2",
            "type": "string",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rating/{book_id}": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Rates a book (1 - 5).",
        "tags": [
          "Rating"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "description": "Book id e.g, '5da1b78a6c649900174252e8'",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "Rating",
            "description": "Enter rating value.",
            "schema": {
              "type": "object",
              "required": [
                "rating_num"
              ],
              "properties": {
                "rating_num": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}